Bugfix/weilei/workspace manager (#157)
* bugfix: list user

* add log

* bugfix: list user failed by system-admin

* bugfix: auth get user

* change the authorization logic for workspace admins.

* add debug log

* supports list users when workspace is nil

* remove debug log

+++ SaFE/bootstrap/upgrade.sh
@@ -53,6 +53,7 @@ echo "✅ Ingress Name: \"$ingress\""
 if [[ "$ingress" == "higress" ]]; then
   echo "✅ Cluster Name: \"$sub_domain\""
 fi
+echo "✅ Upgrade node-agent: \"$install_node_agent\""
 
 echo
 
@@ -125,26 +126,28 @@ install_or_upgrade_helm_chart "$chart_name" "$values_yaml"
 install_or_upgrade_helm_chart "primus-safe-cr" "$values_yaml"
 rm -f "$values_yaml"
 
-echo
-echo "========================================="
-echo "🔧 Step 3: upgrade primus-safe data plane"
-echo "========================================="
+if [[ "$install_node_agent" == "y" ]]; then
+  echo
+  echo "========================================="
+  echo "🔧 Step 3: upgrade primus-safe data plane"
+  echo "========================================="
 
-cd ../node-agent/charts/
-src_values_yaml="node-agent/values.yaml"
-if [ ! -f "$src_values_yaml" ]; then
-  echo "Error: $src_values_yaml does not exist"
-  exit 1
+  cd ../node-agent/charts/
+  src_values_yaml="node-agent/values.yaml"
+  if [ ! -f "$src_values_yaml" ]; then
+    echo "Error: $src_values_yaml does not exist"
+    exit 1
+  fi
+  values_yaml="node-agent/.values.yaml"
+  cp "$src_values_yaml" "${values_yaml}"
+
+  sed -i "s/nccl_socket_ifname: \".*\"/nccl_socket_ifname: \"$ethernet_nic\"/" "$values_yaml"
+  sed -i "s/nccl_ib_hca: \".*\"/nccl_ib_hca: \"$rdma_nic\"/" "$values_yaml"
+  sed -i "s/image_pull_secret: \".*\"/image_pull_secret: \"$IMAGE_PULL_SECRET\"/" "$values_yaml"
+
+  install_or_upgrade_helm_chart "node-agent" "$values_yaml"
+  rm -f "$values_yaml"
 fi
-values_yaml="node-agent/.values.yaml"
-cp "$src_values_yaml" "${values_yaml}"
-
-sed -i "s/nccl_socket_ifname: \".*\"/nccl_socket_ifname: \"$ethernet_nic\"/" "$values_yaml"
-sed -i "s/nccl_ib_hca: \".*\"/nccl_ib_hca: \"$rdma_nic\"/" "$values_yaml"
-sed -i "s/image_pull_secret: \".*\"/image_pull_secret: \"$IMAGE_PULL_SECRET\"/" "$values_yaml"
-
-install_or_upgrade_helm_chart "node-agent" "$values_yaml"
-rm -f "$values_yaml"
 
 echo
 echo "==============================="
